unab-stand-greeter/requirements.txt
opencv-python
mediapipe
numpy
Pillow
pyyaml
pyttsx3
qrcode
